pyproject.toml
[tool.poetry]
name = "gha-changelog-syncer"
version = "0.1.0"
description = "Synchronize GitHub Pull Request to Notion Database"
authors = ["Charly Laurent <charly.laurent@mergify.com>"]
readme = "README.md"
packages = [{include = "gha_changelog_syncer"}]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.23.0"
python-dateutil = "^2.8.2"
daiquiri = "^3.2.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
mypy = "^0.982"
black = {version = "^22.10.0", allow-prereleases = true}
python-dotenv = "^0.21.0"
types-python-dateutil = "^2.8.19"
isort = "^5.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli_handler = 'gha_changelog_syncer.cli_handler:handle'
gha_handler = 'gha_changelog_syncer.gha_handler:handle'
